moved cluster_confidence from utils to streamline per @arokem's suggestion
