+++ dipy/tracking/tests/test_utils.py
@@ -13,8 +13,7 @@ from dipy.tracking.utils import (affine_for_trackvis, connectivity_matrix,
                                  random_seeds_from_mask, target,
                                  target_line_based, unique_rows, near_roi,
                                  reduce_rois, path_length, flexi_tvis_affine,
-                                 get_flexi_tvis_affine, _min_at,
-                                 cluster_confidence)
+                                 get_flexi_tvis_affine, _min_at)
 
 from dipy.tracking.streamline import Streamlines
 
@@ -38,80 +37,6 @@ def make_streamlines():
                              [5, 20, 33],
                              [40, 80, 120]], 'float')]
     return streamlines
-
-
-def test_cluster_confidence():
-    # two identical streamlines should raise an error
-    mysl = np.array([np.arange(10)] * 3, 'float').T
-    test_streamlines = Streamlines()
-    test_streamlines.append(mysl, cache_build=True)
-    test_streamlines.append(mysl)
-    test_streamlines.finalize_append()
-    assert_raises(ValueError, cluster_confidence, test_streamlines)
-
-    # 3 offset collinear streamlines
-    test_streamlines = Streamlines()
-    test_streamlines.append(mysl, cache_build=True)
-    test_streamlines.append(mysl+1)
-    test_streamlines.append(mysl+2)
-    test_streamlines.finalize_append()
-    cci = cluster_confidence(test_streamlines)
-    assert_equal(cci[0], cci[2])
-    assert_true(cci[1] > cci[0])
-
-    # 3 parallel streamlines
-    mysl = np.zeros([10, 3])
-    mysl[:, 0] = np.arange(10)
-    mysl2 = mysl.copy()
-    mysl2[:, 1] = 1
-    mysl3 = mysl.copy()
-    mysl3[:, 1] = 2
-    mysl4 = mysl.copy()
-    mysl4[:, 1] = 4
-    mysl5 = mysl.copy()
-    mysl5[:, 1] = 5000
-
-    test_streamlines_p1 = Streamlines()
-    test_streamlines_p1.append(mysl, cache_build=True)
-    test_streamlines_p1.append(mysl2)
-    test_streamlines_p1.append(mysl3)
-    test_streamlines_p1.finalize_append()
-    test_streamlines_p2 = Streamlines()
-    test_streamlines_p2.append(mysl, cache_build=True)
-    test_streamlines_p2.append(mysl3)
-    test_streamlines_p2.append(mysl4)
-    test_streamlines_p2.finalize_append()
-    test_streamlines_p3 = Streamlines()
-    test_streamlines_p3.append(mysl, cache_build=True)
-    test_streamlines_p3.append(mysl2)
-    test_streamlines_p3.append(mysl3)
-    test_streamlines_p3.append(mysl5)
-    test_streamlines_p3.finalize_append()
-
-    cci_p1 = cluster_confidence(test_streamlines_p1)
-    cci_p2 = cluster_confidence(test_streamlines_p2)
-
-    # test relative distance
-    assert_array_equal(cci_p1, cci_p2*2)
-
-    # test simple cci calculation
-    expected_p1 = np.array([1/1+1/2, 1/1+1/1, 1/1+1/2])
-    expected_p2 = np.array([1/2+1/4, 1/2+1/2, 1/2+1/4])
-    assert_array_equal(expected_p1, cci_p1)
-    assert_array_equal(expected_p2, cci_p2)
-
-    # test power variable calculation (dropoff with distance)
-    cci_p1_pow2 = cluster_confidence(test_streamlines_p1, power=2)
-    expected_p1_pow2 = np.array([np.power(1/1, 2)+np.power(1/2, 2),
-                                 np.power(1/1, 2)+np.power(1/1, 2),
-                                 np.power(1/1, 2)+np.power(1/2, 2)])
-
-    assert_array_equal(cci_p1_pow2, expected_p1_pow2)
-
-    # test max distance (ignore distant sls)
-    cci_dist = cluster_confidence(test_streamlines_p3, max_mdf=5)
-    expected_cci_dist = np.concatenate([cci_p1, np.zeros(1)])
-    assert_array_equal(cci_dist, expected_cci_dist)
 
 
 def test_density_map():
